Moved tests to the right place; added sample serialised twitter graph
