+++ src/test/java/com/bio4j/angulillos/TwitterGraph.java
@@ -1,6 +1,5 @@
-package com.bio4j.angulillos.test;
+package com.bio4j.angulillos;
 
-import com.bio4j.angulillos.*;
 import com.bio4j.angulillos.TypedEdge.Type.*;
 
 public abstract class TwitterGraph <
